backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
httpx>=0.26.0
clerk-backend-api>=1.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
